Fix subfolder handling. Closes #4715
